File for the Design
This is the complete code for the module responsible for
getting input pixel data, handling it to sobel mathematical
module , getting the output and managing the bidirectional 8-bit
bus in synchronous manner with clock
➢ There is one bidirectional 8-bit port called ‘Bus’ carrying data in
the direction specified by the Bus Data Direction{Input Port}.
➢ The other input ports for this module:-
• strobe_data
• strobe_mode
• clock
• reset
• bus_con
➢ The module consists of r1,r2,r3 which are 122-byte Memory elements which
are responsible for storing the rows of pixel data which are used further for
processing.
➢  bus_1 and bus_2 are eight-bit input bitstream and output bitstream governed
by bus_con.
➢  Windex and Rindex are respective addresses for writing into memory and
reading from memory.
➢  ‘write’ and ‘read’ are normal internal controls to prevent crossover.
➢  ‘clk’ is for maintaining the process in synchronous manner and ‘reset’ is used
for clearing all memory data and also registers. It can be majorly used to
ready the device for next input image if any prior processing is done.
➢ The sobel instantiation is done for all memory elements present in the
memory
➢  At positive edge of clock and negative edge of reset , the shift ,read and
write operations are followed accordingly.
➢  While reading back the output values , array of output is read from a
temporary space.